Extract slime.$api.fp.world into its own $api-fp-wo module
Move the Sensor/Means world-oriented API out of $api-fp-impure.fifty.ts
into a dedicated loader/$api-fp-wo.fifty.ts, loaded from
$api-Function.fifty.ts. Also reorders the read declaration in
rhino/file/wo.fifty.ts into its own namespace block.
